rust_static_site_generator
