--- 2021/rust_static_site_generator/src/templates/body.rs
@@ -1,12 +1,0 @@
-pub fn render_body(body: &str) -> String {
-    format!(
-        r#"  <body>
-    <nav>
-        <a href="/">Home</a>
-    </nav>
-    <br />
-    {}
-  </body>"#,
-        body
-    )
-}
