--- 2021/rust_static_site_generator/src/templates/footer.rs
@@ -1,4 +1,0 @@
-pub const FOOTER: &str = r#"
-
-</html>
-"#;
