--- 2021/rust_static_site_generator/src/templates/header.rs
@@ -1,9 +1,0 @@
-pub const HEADER: &str = r#"<!DOCTYPE html>
-<html lang="en">
-
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-  </head>
-
-"#;
