--- 2021/rust_static_site_generator/src/templates/mod.rs
@@ -1,7 +1,0 @@
-mod body;
-mod footer;
-mod header;
-
-pub use body::render_body;
-pub use footer::FOOTER;
-pub use header::HEADER;
